Adds initial splat configuration & project structure
Adds a basic splat configuration for the NTSC-U version of the game to get started with the reimplementation process.

Also configures the basic project structure where each version will be contained to it's own folder.

Configures the .gitignore to ensure the binary and output files will not ever be added to the version control during development.