Adds script to remove unneeded object sections to workaround linker errors
This script is an extremely hacky workaround for a linker error (e.g. ".bss(.bss) in file <file>.s.o is referenced but has not". The linker is upset with sections in the .o files that are empty when linking even though they are expected to be.

This script is being used to remove those sections to make the linker content. It is currently unclear for a better way to handle this as it is also unclear why this project is experiencing this issue.