Splat: Mark all other segments and set elf_header as bin
Creates segments for the other sections of the elf file so that potentially they can be linked in later, but for now just as .bin files since I don't want to include them in my linker yet.

Also setting the header to bin as well as we're not yet linking it in.