Ignore iso and elf output as well as other files for building
Ignores iso files related to building and elf files that are also part of the build process.

Also ignores the binutils folder which will be downloaded as a requirement for the build process (the decompals binutils).

Also ignores a sort.txt file used in the extracting/rebuilding of the iso.