Make: Download MWCC
Adds step to download the needed MW 2.4 compiler from decompme and extracts it into the compilers folder. However, the makefile has not yet been updated to change paths to use the downloaded version. This is included as a convenience to be later integrated. For now it is just here for manual use.