Adds better setup documentation for WSL & dependencies + venv setup target
Adds a target for make that includes the commands for configuring venv to make it more convenient for a fresh setup.

Adds updated documentation to the readme for dependencies.
Adds a setup guide for Ubuntu WSL.
Adds a minor comment in the make file explaining what the fix linker paths script is for.