Splat: Correcting alignment and size of data/sdata/sbss/bss symbols
Updates the sizes and sometimes types of the symbols for the data/sdata/sbss/bss symbol sections to improve the data match to the original elf file.

Fixes the positions of a few splat splits to correctly align to the symbol start and ends.

Adds additional comments to help make the symbols easier to identify and organize.

Sets a number of MWLD linker symbols to ignore since they will be generated at link time.

Removes from alignment directives in the linker script to further correct the alignment of the sections.