Make: Add downloading coddog & configure requirements-dev.txt
Coddog is used for finding matching functions and other functionality. This is an optional dev dependency.  It's not required for building the project, simply for developing the decomp it's a useful utility.

Also configures a 'requirements-dev.txt' to include optional dev only python dependencies.

Configures a new make target 'make install-dev' to install optional dev dependencies via python and other optional utilities like coddog.

Updates readme to account for the change in dependencies installation.